pyproject.toml
[project]
name = "generated-adventures"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "pydantic-ai>=0.6.2",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "nicegui>=1.4.0",
    "uvicorn>=0.35.0",
    "google-generativeai>=0.7.2",
    "Pillow>=10.5.0",
]
